SerialServer: Fix switching instability
If the serial server received two \n and/or \r characters in a row,
then it would reset back to the beginning of the FSM. This resulted
in needing a certain timing to successfully switch VMs.

This change fixes this behavior by staying in the same state when
multiple \n and/or \r characters are received.

Change-Id: I29df5a8f7cd1107fd7b49a49d3698df08a6c982a